% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_duplicate_dfs.R
\name{combine_duplicate_dfs}
\alias{combine_duplicate_dfs}
\title{Combine data frames with the same name}
\usage{
combine_duplicate_dfs(df_list)
}
\arguments{
\item{df_list}{A list of data frames.}
}
\value{
A list of data frames whose length is shorter than that of the input list.

The length of the output list can be determined by \code{n - d + u}, where \code{n} is the length of the original list, \code{d}
is the number of duplicate names in the list, and \code{u} is the number of unique duplicate names in the list.
}
\description{
\code{combine_duplicate_dfs()} takes a list of data frames as input and combines data frames with identical names in the
list into unified data frame under the same name. Partitioned files that are read into data frames will have
identical names, so \code{combine_duplicate_dfs()} essentially de-partitions those data.
}
\examples{
# Create some example data frames
df1 <- data.frame(x = 1:3, y = letters[1:3])
df2 <- data.frame(x = 4:6, y = letters[4:6])
df3 <- data.frame(x = 7:9, y = letters[7:9])
df4 <- data.frame(x = 10:12, y = letters[10:12])
df_list <- list(df1, df2, df1, df3, df4, df2)
names(df_list) <- c("df1", "df2", "df1", "df3", "df4", "df2")

# Combine the data frames
combined_df_list <- combine_duplicate_dfs(df_list)
head(combined_df_list)
}
