% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_col_to_numeric.R
\name{convert_col_to_numeric}
\alias{convert_col_to_numeric}
\title{Convert a specific column of specific data frames to numeric type}
\usage{
convert_col_to_numeric(
  df_list,
  df_to_avoid = "device",
  col_to_convert = "value"
)
}
\arguments{
\item{df_list}{A list of data frames.}

\item{df_to_avoid}{A string denoting the full or partial name of a data frame in \code{df_list}.}

\item{col_to_convert}{The name of a column to convert to numeric type.}
}
\value{
A list of data frames with modified column types and is the same size as \code{df_list}.
}
\description{
\code{convert_col_to_numeric()} takes a list of data frames, a string denoting the full or partial name of a
data frame, and a column name as input, then converts all instances of the specified column (\code{col_to_convert}) in
the data frames in \code{df_list} to numeric type, except in the data frames whose names contain the string specified by
\code{df_to_avoid}.
}
\examples{
# Example 1: Convert "value" column in all data frames of list `my_dfs` to numeric type
my_dfs <- list(data.frame(value = c("1", "2", "3")),
               data.frame(value = c("4", "5", "6")),
               data.frame(value = c("7", "8", "9")))

names(my_dfs) <- c("1", "2", "3")

my_dfs <- convert_col_to_numeric(my_dfs)

# Example 2: Convert "my_col" column in all data frames of 
# my_list to numeric type, except in "exclude_df"
my_list <- list(df1 = data.frame(my_col = c("1", "2", "3")),
                df2 = data.frame(my_col = c("4", "5", "6")),
                exclude_df = data.frame(my_col = c("7", "8", "9")))

my_list <- convert_col_to_numeric(my_list, "exclude", "my_col")
}
