% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_concepts.R
\name{diff_concepts}
\alias{diff_concepts}
\title{Return the set difference of column names between data frames}
\usage{
diff_concepts(df_list, concept_replacements, concept_map, concept_filter_col)
}
\arguments{
\item{df_list}{A list of data frames.}

\item{concept_replacements}{A named vector (names must be valid values of \code{concept_filter_col} column of
\code{concept_map} data frame).}

\item{concept_map}{A data frame (for i2b2/RECOVER purposes, this data frame is the ontology file).}

\item{concept_filter_col}{The column of \code{concept_map} that contains "approved concepts" (column names of data frames
in \code{df_list} that are not to be excluded).}
}
\value{
A vector of column names from the data frames in \code{df_list} that are not present in the values of the
\code{concept_filter_col} column of \code{concept_map}.
}
\description{
\code{diff_concepts()} takes a list of data frames and a secondary data frame (\code{concept_map}) as input, along with a named
vector and the name of a column in \code{concept_map}. First, \code{diff_concepts()} replaces various strings in
\code{concept_filter_col} with the named values in \code{concept_replacements}. Then, it returns a vector of column names that
are present in the data frames of \code{df_list} and not present in the values of the specified column
(\code{concept_filter_col}) in \code{concept_map}.
}
\examples{
# Create a sample data frame
first1 <- data.frame(col_a = 1:5, col_b = letters[1:5], col_c = TRUE)

# Create another sample data frame
second2 <- data.frame(col_b = letters[6:10], col_d = 6:10, col_e = FALSE)

# Create a list of the data frames
df_list <- list(first1, second2)

# Set the names of the data frames in `df_list`
names(df_list) <- c("first1", "second2")

# Create a `concept_map` data frame
concept_map <- data.frame(concept = c("summary:a", "summary:b", "trigger:c"))

# Define the concept replacements
concept_replacements <- c("a" = "col_a", "b" = "col_b")

# Get the set difference of column names between data frames
diff_concepts(df_list = df_list,
              concept_replacements = concept_replacements,
              concept_map = concept_map,
              concept_filter_col = "concept")
}
