% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concept_map.R
\name{get_concept_map}
\alias{get_concept_map}
\title{Fetch Synapse CSV file and store in data frame}
\usage{
get_concept_map(synID)
}
\arguments{
\item{synID}{A Synapse ID for a CSV file in Synapse.}
}
\value{
A data frame containing a representation of the data in the file. The values in the "concept_cd" column are
converted to lowercase.
}
\description{
\code{get_concept_map()} fetches the concept map (ontology) CSV file from Synapse based on the synID of the stored file,
then reads the file to a data frame.
}
\examples{
\dontrun{
df <- get_concept_map("syn12345678")
}

}
