% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_df.R
\name{melt_df}
\alias{melt_df}
\title{Melt specific columns of a data frame}
\usage{
melt_df(df, excluded_concepts = "")
}
\arguments{
\item{df}{A data frame.}

\item{excluded_concepts}{A vector of column names to not melt.}
}
\value{
A data frame of the same type as \code{df} but different shape.
}
\description{
\code{melt_df()} takes a data frame and a vector as input, and reshapes only the specified columns of the data
frame from wide to long format. While keeping the columns whose names are in the "excluded concepts" vector untouched,
\code{melt_df()} reshapes only the remaining columns of \code{df}.

The resulting data frame has a "concept" column containing the melted column names (those that are not found in
\verb{excluded concepts}), and a "value" column containing the values of those melted columns.
}
\examples{
#' # Example 1: Melt all columns except "Species" of the "iris" dataset
head(iris)
melted_df <- melt_df(iris, c("Species"))
head(melted_df)

# Example 2: Melt columns "col1" and "col2" of a custom data frame
df <- data.frame(col1 = c(1,2,3), col2 = c(4,5,6), col3 = c(7,8,9), col4 = c(10, 11, 12))
head(df)
melted_df <- melt_df(df, c("col3", "col4"))
head(melted_df)

# Example 3: Melt all columns of the "mtcars" dataset
head(mtcars)
melted_df <- melt_df(mtcars)
head(melted_df)
}
