% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_str_pairs.R
\name{reverse_str_pairs}
\alias{reverse_str_pairs}
\title{Reverse order of a named vector}
\usage{
reverse_str_pairs(str_pairs)
}
\arguments{
\item{str_pairs}{A named vector.}
}
\value{
A named vector of the same type as \code{str_pairs}.
}
\description{
\code{reverse_str_pairs()} takes a named vector as input and returns a new named vector, where both the order of the input vector and its names and values are reversed.
}
\examples{
# Create a named vector of strings
fruit_colors <- c(apple = "red", banana = "yellow", cherry = "red",
                  kiwi = "brown", lemon = "yellow")

# Reverse the order of the named vector and its values
reversed_fruit_colors <- reverse_str_pairs(fruit_colors)

# Print the original and reversed named vectors
print(fruit_colors)
print(reversed_fruit_colors)
}
