% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_in_syn.R
\name{store_in_syn}
\alias{store_in_syn}
\title{Store Files in Synapse}
\usage{
store_in_syn(synFolderID, filepath, used_param = NULL, executed_param = NULL)
}
\arguments{
\item{synFolderID}{A Synapse ID for a folder entity in Synapse.}

\item{filepath}{The path for the file to be uploaded to Synapse.}

\item{used_param}{The Synapse ID, URL, file name (if in the same working directory), or file path used to create the
object (can also be a list of these).}

\item{executed_param}{The Synapse ID, URL, file name (if in the same working directory), or file path representing
code executed to create the object (can also be a list of these).}
}
\value{
A Synapse Entity.
}
\description{
\code{store_in_syn()} creates a new Synapse Entity with any associated provenance, if provided, uploading any files in the
process.
}
\examples{
\dontrun{
synFolderID <- "syn12345678"
ontologyFileID <- "syn87654321"

store_in_syn(synFolderID,
             'file1.csv',
             used_param = ontologyFileID,
             executed_param = "https://github.com/username/repo/blob/branch/script.R")
             
store_in_syn(synFolderID,
             'file2.csv',
             used_param = ontologyFileID)
}
}
