% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pipeline.R
\name{summarize_pipeline}
\alias{summarize_pipeline}
\title{Execute recoverSummarizeR main pipeline}
\usage{
summarize_pipeline(
  ontologyFileID,
  parquetDirID,
  dataset_name_filter,
  concept_replacements,
  concept_filter_col,
  synFolderID,
  method = "synapse",
  s3bucket = NULL,
  s3basekey = NULL,
  downloadLocation
)
}
\arguments{
\item{ontologyFileID}{A Synapse ID for a CSV file stored in Synapse. For RECOVER, this file is the i2b2 concepts map.}

\item{parquetDirID}{A Synapse ID for a folder entity in Synapse where the data is stored. For RECOVER, this would be
the folder housing the post-ETL parquet data.}

\item{dataset_name_filter}{A string used to filter the list of data frames by including only the data frames whose
names contain the specified string.}

\item{concept_replacements}{A named vector (names must be valid values of \code{concept_filter_col} column of
\code{concept_map} data frame).}

\item{concept_filter_col}{The column of \code{concept_map} that contains "approved concepts" (column names of data frames
in \code{df_list} that are not to be excluded).}

\item{synFolderID}{A Synapse ID for a folder entity in Synapse.}

\item{method}{Either 'synapse' or 'sts' to specify the method to use in getting the parquet datasets. 'synapse' will get files directly from a synapse project or folder using the synapse client, while 'sts' will use sts-token access to get objects from an sts-enabled storage location, such as an S3 bucket.}

\item{s3bucket}{The name of the S3 bucket to access when \code{method='sts'}.}

\item{s3basekey}{The base key of the S3 bucket to access when \code{method='sts'}.}

\item{downloadLocation}{The location to download files to.}
}
\description{
Execute recoverSummarizeR main pipeline
}
\examples{
\dontrun{
synapser::synLogin()
ontologyFileID <- "syn12345678"
parquetDirID <- "syn87654321"
dataset_name_filter <- "fitbit"
concept_replacements <- c("mins" = "minutes",
                          "avghr" = "averageheartrate",
                          "spo2" = "spo2_",
                          "hrv" = "hrv_dailyrmssd",
                          "restinghr" = "restingheartrate",
                          "sleepbrth" = "sleepsummarybreath")
concept_filter_col <- "concept_cd"
synFolderID <- "syn18273645"
method <- "synapse"
s3bucket <- "my-project-bucket"
s3basekey <- "main/parquet/"
downloadLocation <- "./parquet"

summarize_pipeline(ontologyFileID = ontologyFileID,
                   parquetDirID = parquetDirID,
                   dataset_name_filter = dataset_name_filter,
                   concept_replacements = concept_replacements,
                   concept_filter_col = concept_filter_col,
                   synFolderID = synFolderID,
                   method = method,
                   s3bucket = s3bucket,
                   s3basekey = s3basekey,
                   downloadLocation = downloadLocation)
}
}
