% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synget_file_to_df.R
\name{syn_file_to_df}
\alias{syn_file_to_df}
\title{Fetch Synapse CSV file and store in data frame}
\usage{
syn_file_to_df(synID)
}
\arguments{
\item{synID}{A Synapse ID for a CSV file in Synapse.}
}
\value{
A data frame containing a representation of the data in the file. The values in the "concept_cd" column are
converted to lowercase.
}
\description{
\code{syn_file_to_df()} fetches the concept map (ontology) CSV file from Synapse based on the synID of the stored file,
then reads the file to a data frame.
}
\examples{
\dontrun{
df <- syn_file_to_df("syn12345678")
}

}
