% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_parquet_dataset_to_dflist.R
\name{syn_parquet_dataset_to_dflist}
\alias{syn_parquet_dataset_to_dflist}
\title{Download Parquet Datasets from Synapse or an STS-Enabled Location}
\usage{
syn_parquet_dataset_to_dflist(
  synDirID,
  method = "synapse",
  s3bucket = NULL,
  s3basekey = NULL,
  downloadLocation,
  dataset_name_filter = NULL
)
}
\arguments{
\item{synDirID}{A Synapse ID for a folder entity in Synapse.}

\item{method}{Either 'synapse' or 'sts' to specify the method to use in getting the parquet datasets. 'synapse' will get files directly from a synapse project or folder using the synapse client, while 'sts' will use sts-token access to get objects from an sts-enabled storage location, such as an S3 bucket.}

\item{s3bucket}{The name of the S3 bucket to access when \code{method='sts'}.}

\item{s3basekey}{The base key of the S3 bucket to access when \code{method='sts'}.}

\item{downloadLocation}{The location to download files to.}

\item{dataset_name_filter}{A string used to filter the list of data frames by including only the data frames whose
names contain the specified string.}
}
\value{
A list of data frames. The list is filtered to include only the data frames whose names contain the string
specified by \code{dataset_name_filter}.
}
\description{
\code{syn_parquet_dataset_to_dflist()} downloads objects from a Synapse directory or syncs from an STS-enabled location, such as an S3 bucket, and maintains the original directory
structure, reads the parquet datasets to data frames, and stores the data frames in a list.
}
\examples{
\dontrun{
df_list <- syn_parquet_dataset_to_dflist("syn12345678", 
                                         "sts", 
                                         "s3://my-bucket/", 
                                         "main/folder", 
                                         "./parquet", 
                                         "fitbit")
# df_list will use the `sts` method to get objects in synapse at 
  the synDirID and the output will contain only data frames for 
  parquet datasets whose names contain the string "fitbit"
}
}
