% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_reverse.R
\name{vec_reverse}
\alias{vec_reverse}
\title{Reverse the order and pairings of a named vector}
\usage{
vec_reverse(str_pairs)
}
\arguments{
\item{str_pairs}{A named vector.}
}
\value{
A named vector of the same type as \code{str_pairs}.
}
\description{
\code{vec_reverse()} reverses both the order of a named vector and its name = value pairings.
}
\examples{
# Create a named vector of strings
fruit_colors <- c(apple = "red", banana = "yellow", cherry = "red",
                  kiwi = "brown", lemon = "yellow")

# Reverse the order of the named vector and its values
reversed_fruit_colors <- vec_reverse(fruit_colors)

# Print the original and reversed named vectors
print(fruit_colors)
print(reversed_fruit_colors)
}
