% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_df.R
\name{melt_df}
\alias{melt_df}
\title{Reshape selected columns from wide to long format}
\usage{
melt_df(df, excluded_concepts = "")
}
\arguments{
\item{df}{A data frame.}

\item{excluded_concepts}{A vector of column names to not melt.}
}
\value{
A data frame of the same type as \code{df} but different shape. The resulting data frame has two new columns:
"concept," which contains the names of the melted columns, and "value," which contains the values of those melted
columns. \code{melt_df()} does not modify the excluded columns and only reshapes the remaining columns.
}
\description{
\code{melt_df()} reshapes specific columns of a given data frame from wide to long format. The "excluded concepts" vector
allows users to specify which columns should be excluded from this reshaping.
}
\examples{
#' # Example 1: Melt all columns except "Species" of the "iris" dataset
head(iris)
melted_df <- melt_df(iris, c("Species"))
head(melted_df)

# Example 2: Melt columns "col1" and "col2" of a custom data frame
df <- data.frame(col1 = c(1,2,3), col2 = c(4,5,6), col3 = c(7,8,9), col4 = c(10, 11, 12))
head(df)
melted_df <- melt_df(df, c("col3", "col4"))
head(melted_df)

# Example 3: Melt all columns of the "mtcars" dataset
head(mtcars)
melted_df <- melt_df(mtcars)
head(melted_df)
}
