% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_df.R
\name{process_df}
\alias{process_df}
\title{Process a data frame with concept mappings}
\usage{
process_df(
  df,
  concept_map,
  concept_replacements_reversed,
  concept_map_concepts = "concept_cd",
  concept_map_units = "UNITS_CD"
)
}
\arguments{
\item{df}{A data frame with \code{participantidentifier}, \code{startdate}, \code{enddate}, \code{concept}, and \code{value} columns.}

\item{concept_map}{A data frame (for i2b2/RECOVER purposes, this data frame is the ontology file).}

\item{concept_replacements_reversed}{A lowercase named vector that can be created using
\code{\link{reverse_str_pairs}}; names must be valid values of \code{df$concept}, and values must contain the original
names from the \code{concept_replacements} named vector used for \code{\link{diff_concepts}}}

\item{concept_map_concepts}{The name of the column in \code{concept_map} that is the equivalent of \code{df$concept}.}

\item{concept_map_units}{The name of the column in \code{concept_map} containing values/info for the units of each value
in \code{df$concept}.}
}
\value{
A data frame with a different shape than that of \code{df}.
}
\description{
\code{process_df()} takes a primary data frame (\code{df}), a secondary data frame (\code{concept_map}), and named
vector as inputs. First, the modifications made to the strings in the \code{concept_filter_col} column in \code{concept_map} by
passing \code{concept_replacements} to \code{\link{diff_concepts}} are reversed using the \code{name = value} pairs from
\code{concept_replacements_reversed}. Then, various columns are added to \code{df}, and the modified data frame is returned. An
error is raised if a specific set of columns are not found in \code{df}.
}
\examples{
# Create a sample data frame
df <- data.frame(participantidentifier = c(1, 2, 3, 4),
                 startdate = c("2022-01-01", "2022-01-01", "2022-01-02", "2022-01-03"),
                 enddate = c("2022-01-03", "2022-01-02", "2022-01-03", "2022-01-04"),
                 concept = c("summary:BLOOD_PRESSURE_SYSTOLIC", "summary:BLOOD_PRESSURE_DIASTOLIC",
                             "summary:HEART_RATE", "summary:RESPIRATORY_RATE"),
                 value = c(120, 80, 70, 20))
print(df)

# Create a sample `concept_map` data frame
concept_map <- data.frame(concept_cd = c("summary:BP_SYSTOLIC", "summary:BP_DIASTOLIC",
                                         "summary:HR", "summary:RR"),
                          valtype_cd = c("N", "N", "N", "N"),
                          UNITS_CD = c("mmHg", "mmHg", "beats/min", "breaths/min"))
print(concept_map)

# Assume `concept_replacements` already exists
concept_replacements <- c("bp" = "blood_pressure", "hr" = "heart_rate", "rr" = "respiratory_rate")

# Create a reversed named vector of concept replacements
concept_replacements_reversed <- reverse_str_pairs(concept_replacements)
print(concept_replacements_reversed)

# Process the data frame
df_processed <- process_df(df, concept_map, concept_replacements_reversed, "concept_cd", "UNITS_CD")
head(df_processed)
}
