% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summarize.R
\name{stat_summarize}
\alias{stat_summarize}
\title{Summarize data in a data frame for i2b2/RECOVER}
\usage{
stat_summarize(df)
}
\arguments{
\item{df}{A data frame with \code{participantidentifier}, \code{concept}, \code{value}, and some combination of \code{startdate},
\code{enddate}, \code{date}, and \code{datetime} columns.}
}
\value{
A data frame that differs in size from \code{df}.
}
\description{
\code{stat_summarize()} takes a data frame as input and summarizes the data in the data frame on specific time scales
(weekly and all-time) for the following statistics: 5/95 percentiles, mean, median, variance, number of records.
}
\examples{
# Create sample data
df <- data.frame(participantidentifier = c(rep("A", 6), rep("B", 6)),
                 date = c("2022-01-01", "2022-01-05", "2022-01-10", "2022-02-01",
                          "2022-02-05", "2022-02-10", "2022-01-01", "2022-01-06",
                          "2022-01-11", "2022-02-01", "2022-02-06", "2022-02-11"),
                 concept = c("weight", "weight", "weight", "weight", "weight", "weight",
                             "height", "height", "height", "height", "height", "height"),
                             value = c(60, 62, 64, 65, 66, 68, 160, 162, 164, 165, 166, 168))
# Summarize the data
stat_summarize(df)
}
