% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_file_to_df.R
\name{syn_file_to_df}
\alias{syn_file_to_df}
\title{Fetch a Synapse file and store in data frame}
\usage{
syn_file_to_df(synID, cols = "concept_cd")
}
\arguments{
\item{synID}{A Synapse ID for a file in Synapse.}
}
\value{
A data frame containing a representation of the data in the file. The values in the columns specified by
\code{cols} are converted to lowercase. For i2b2/RECOVER, the default has been set to \code{concept_cd}.
}
\description{
\code{syn_file_to_df()} fetches a file from Synapse based on the synID of the stored file, then reads the file to a data
frame, with the option to lowercase a specific column. For i2b2/RECOVER, \code{syn_file_to_df()} can be used to fetch the
concept map (ontology) file from Synapse then read it into a data frame.
}
\examples{
\dontrun{
df <- syn_file_to_df("syn12345678")
}

}
